KVM: PPC: Book3S HV P9: implement hash host / hash guest support
This additionally has to save and restore the host SLB, and also
ensure that the MMU is off while switching into the guest SLB.

P9 and later CPUs now always go via the P9 path. The "fast" guest
mode is now renamed to the P9 mode, which is consistent with
functionality and naming.

Signed-off-by: Nicholas Piggin <npiggin@gmail.com>