KVM: PPC: Book3S HV Nested: Indicate guest PMU in-use in VPA
After saving the host PMU SPRs and before loading the guest's, switch
the pmcregs_in_use field in our VPA (lppaca) to the value advertised by
the guest in its VPA.

This passes through the PMU status from the guest to the parent
hypervisor, so it will save and restore the PMU as necessary when
exiting from the nested guest (or the region when we have the guest
PMU SPRs loaded).

This fixes the nested HV bug for which a workaround was added to the L0
/ parent HV by commit 63279eeb7f93a ("KVM: PPC: Book3S HV: Always save
guest pmu for guest capable of nesting"), which explains the problem in
detail. The workaround is no longer required in the parent HV for a
nested capable guest with this fix.

Fixes: 95a6432ce903 ("KVM: PPC: Book3S HV: Streamlined guest entry/exit path on P9 for radix guests")
Signed-off-by: Nicholas Piggin <npiggin@gmail.com>