KVM: PPC: Book3S HV P9: Move PMU switching into its own code
This moves PMU switching out into its own routines and moves it
right up to the point before asm is called into, which helps the
host perf get the most coverage of KVM guest entry/exit code.

This is slightly suboptimal for SPR scheduling point of view when
the PMU is enabled, but when it's disabled there is no difference.