KVM: PPC: Book3S HV: Don't always save PMU for guest capable of nesting
Revert the workaround added by commit 63279eeb7f93a ("KVM: PPC: Book3S
HV: Always save guest pmu for guest capable of nesting"). Nested capable
guests running with the earlier commit ("KVM: PPC: Book3S HV Nested:
Indicate guest PMU in-use in VPA") now indicate the PMU in-use status of
their guests, which means the PMU does not need to be saved if it's
marked not in use in the guest VPA.

This will cause perf breakage for older nested guests without that fix.
It's unclear there's an easy way to avoid that, so this could wait for a
release or so for the fix to filter into stable kernels.

Signed-off-by: Nicholas Piggin <npiggin@gmail.com>