powerpc/pseries/idle: reduce priority nops in snooze loop
Rather than hammering priority nops in the snooze loop, set it once to
very low and spin with that (also add a cpu-relax). This can cause an
interrupt to come in and raise priority but not break out of idle which
will mean spinning continues at normal priority, but this should be
relatively uncommon and we break out soon enough anyway. A complete
patch would either change how PPR is managed, or hold interrupts disabled
in the idle loop and check for pending interrupt.