KVM: PPC: Book3S: Fix H_RTAS rets buffer overflow
The kvmppc_rtas_hcall sets the host rtas_args.rets pointer based on the
rtas_args.nargs that was provided by the guest. The guest nargs value is
not range checked, so the guest can cause the host rets pointer to be set
outside the args array. The individual rtas function handlers check the
nargs and nrets values to ensure they are correct, but in case they are
not, the handlers store a -3 failure indication in rets[0] which corrupts
host memory.

Fix this by testing up front whether the guest supplied nargs and nret
would exceed the array size, and fail the hcall directly without storing
a failure indication to rets[0].

Also expand on a comment about why we kill the guest and try not to return
errors directly if we have a valid rets[0] pointer.

Reported-by: Alexey Kardashevskiy <aik@ozlabs.ru>
Fixes: 8e591cb720473 (KVM: PPC: Book3S: Add infrastructure to implement kernel-side RTAS calls")
Signed-off-by: Nicholas Piggin <npiggin@gmail.com>