locking/qspinlock: shrink struct qnode so all qnodes fit in a 64-byte cacheline
count and locked can be u8. The concern over cacheline usage here is
overblown though (I mean it's great, but using one additional cache
line in the case you have a contended process-context spinlock that's
interrupted by a contended bh spinlock that's interrupted by a contended
irq spinlock isn't going to be the end of the world).

Signed-off-by: Nicholas Piggin <npiggin@gmail.com>