powerpc/qspinlock: allow new waiters to steal the lock before queueing
Allow new waiters a number of spins on the lock word before queueing,
which particularly helps paravirt performance when physical CPUs are
oversubscribed.

The new __queued_spin_trylock_steal() function is put in qspinlock.h
because it will be used there in a later change, so save having to
move it.

Signed-off-by: Nicholas Piggin <npiggin@gmail.com>