KVM: PPC: selftests: add support for powerpc
Implement KVM selftests support for powerpc (Book3S-64).

ucalls are implemented with an unsuppored PAPR hcall number which causes
KVM to exit to userspace.

For now virtual memory is only implemented for the radix MMU, and only
the base page size is supported.

Guest interrupts are taken in real-mode, so require a real-mode page at
gRA 0. Interrupt entry requires some tricky code because gVA:gRA is not
1:1 mapped like the kernel is, so we can't just switch MMU on and off.

Acked-by: Michael Ellerman <mpe@ellerman.id.au> (powerpc)
Signed-off-by: Nicholas Piggin <npiggin@gmail.com>